CHORE: update pre-commit
Poetry in the pre-commit has being updated to the 2.0 stream. This has
meant the remove of the `poetry-export` hook, and it is no longer
supported. Because of this the `requirements.txt` file has also being
removed.

It would have being too hard to maintain the requirements file on 2.0 of
poetry. Its better not there than wrong. There was also the fact as
there was no external dependencies the file should have being empty.

Signed-off-by: Jim Fitzpatrick <jimfity@gmail.com>